useState and useContext hooks working
useState and useContext hooks working somewhat, need to figure out toggle of third mode
