+++ src/components/header.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import Nav from './nav'
 
+
 const Header = () => {
     return (
         <>
