--- src/styles/theme.js
@@ -1,11 +1,0 @@
-export const lightTheme = {
-    
-}
-
-export const darkTheme = {
-    
-}
-
-export const starTheme = {
-    
-}
